Completed_Add_DElete

+++ ExpenseTrackerBackend/controller/expenseC.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const sequelize = require('sequelize');
 const expenseT = require('../model/expenseT');
 let expenseHtmlFile = path.join(__dirname,'../../ExpenseTrackerFrontEnd/view/expenses/expenseForm.html');
 
@@ -17,26 +18,28 @@ exports.profile =(req, res, next) => {
 exports.addExpense=async (req,res,next)=>{
   try {
 
-    const {amount , description ,category} = req.body;
+    const {amount, description, category} = req.body;
 
    if(amount == undefined || amount === 0){
     return res.status(400).json({success:false});
    }
-
-    const response  = await expenseT.create({amount , description ,category});
+    console.log(`amount is ${amount}  description is ${description}   category is ${category}  user id is ${req.user.id}`);
+    const response = await expenseT.create({amount,description,category, userId: req.user.id});
     console.log(response);
     return res.status(201).json({response, success:true });
-
   } catch(err){
+    console.log(err);
     return res.status(403).json({success:false , error:err});
   }  
 }
 
 exports.getExpense = async(req,res,next)=>{
   try {
-    const expenses  =  await expenseT.findAll();
+    console.log("this is user IDDDDDDDDDDDDDD"+req.user.id);
+    const expenses  = await req.user.getExpenses();//Insted of this magic function works like this const expenses  =  await expenseT.findAll({where:{userId:req.user.id}});
       return res.status(200).json({expenses,success:true});
-  } catch(error){
+  } catch(err){
+    console.log(err)
     return res.status(500).json({error:err,success:false});
   }
 }
@@ -49,7 +52,7 @@ exports.deleteExpense=async(req,res,next)=>{
    if(id==undefined || id.length === 0){
     return res.status(400).json({success:false,message:"trying to delete unavailable item"});
    }
-   await expenseT.destroy({where :{id:id}})
+   await expenseT.destroy({where :{id:id , userId : req.user.id}})
    return res.status(200).json({success:true,message:"Deleted Successfully"});
   } catch (err) {
     return res.status(500).json({error:err,success:false,message:"Failed In Deleting"});
